Update file_spec.h
This arose from a peculiar usage case where mgiza was being called from a PHP script on an Apache server. In this case, it appears that the getenv() call returned a null pointer, which segfaulted when strlen() was called on it.

I'm not sure if the better solution is to return "nobody" or simply an empty string, but either should fix the null pointer problem.